prettier issue fixed.

+++ contracts/oracles/default/UniswapV3PriceOracle.sol
@@ -138,12 +138,16 @@ contract UniswapV3PriceOracle is IPriceOracle {
     return getPriceX96FromSqrtPriceX96(pool.token0(), token, sqrtPriceX96);
   }
 
-  function getPriceX96FromSqrtPriceX96(address token0, address priceToken, uint160 sqrtPriceX96) public pure returns (uint256 price) {
+  function getPriceX96FromSqrtPriceX96(
+    address token0,
+    address priceToken,
+    uint160 sqrtPriceX96
+  ) public pure returns (uint256 price) {
     if (token0 == priceToken) {
-        price = FullMath.mulDiv(sqrtPriceX96, sqrtPriceX96, uint(2**(96*2)) / 1e18);
+      price = FullMath.mulDiv(sqrtPriceX96, sqrtPriceX96, uint256(2**(96 * 2)) / 1e18);
     } else {
-        price = FullMath.mulDiv(sqrtPriceX96, sqrtPriceX96, uint(2**(96*2)) / 1e18);
-        price = 1e36 / price;
+      price = FullMath.mulDiv(sqrtPriceX96, sqrtPriceX96, uint256(2**(96 * 2)) / 1e18);
+      price = 1e36 / price;
     }
   }
 }
